Remove printout from MessageProc()
Apparently it causes race conditions with the new'ing of the output
class, which cause Winamp to crash unless you run it in the debugger.
Since it was only for debugging anyway, and I never figured out if I
needed to do anything with those messages, no point in keeping it.